Split up optoins dialog
