+++ PlattformOrdMan/UI/Dialog/OptionsDialog/OrderHistoryOptionOK.cs
@@ -1,0 +1,4 @@
+﻿namespace PlattformOrdMan.UI.Dialog.OptionsDialog
+{
+    public delegate void OrderHistoryOptionOK(bool isIncludedColumnsUpdated);
+}
