+++ PlattformOrdMan/UI/Dialog/OptionsDialog/OrderHistoryOptionsDialog.resx
